+++ docs/release_notes.rst
@@ -10,6 +10,9 @@ v2.1.0
 * Updated the default args to `SwiftPath.upload`. Static large objects are
   used by default.
 
+* add `SwiftPath.expand`, `SwiftPath.expandvars` and `SwiftPath.normpath`
+  following ``os.path`` versions of functions.
+
 v2.0.0
 ------